tightening

+++ DataProvider2/Models/CustomerTransaction.cs
@@ -7,7 +7,7 @@
         public int TransactionId { get; set; }
         public DateTime TransactionDate { get; set; }
         public string Type { get; set; }
-        public Decimal Value { get; set; }
+        public double Value { get; set; }
         public string CustomerCode { get; set; } = string.Empty;
         public string Name { get; set; } = string.Empty;
